+++ src/memoryValue.h
@@ -11,8 +11,8 @@ class MemoryValue {
     public:
         MemoryValue(int setAddress, int valueAddress);
         MemoryValue(int setAddress, int valueAddress, int defaultValue);
-        void setValue(uint8_t value);
-        uint8_t readValue();
+        void setValue(uint16_t value);
+        uint16_t readValue();
 
         bool isSet();
         void unset();
